Core: move digest-related functions from lib/common/xml.c to new digest.c
Specifically:
 dump_xml_for_digest()
 calculate_xml_digest_v1()
 calculate_xml_digest_v2()
 calculate_on_disk_digest()
 calculate_operation_digest()
 calculate_xml_versioned_digest()

Because this required that dump_xml() and fix_digest_buffer() be non-static,
those were renamed to crm_xml_dump() and crm_buffer_add_char() respectively
and declared in crm_internal.h.
