+++ include/crm_internal.h
@@ -347,5 +347,7 @@ static inline void *realloc_safe(void *ptr, size_t size)
     return ret;
 }
 
+void crm_xml_dump(xmlNode * data, int options, char **buffer, int *offset, int *max, int depth);
+void crm_buffer_add_char(char **buffer, int *offset, int *max, char c);
 
 #endif                          /* CRM_INTERNAL__H */
